Prefer project systemjs-bundle over /usr/local/bin/systemjs-bundle
Because `->defaultValue('/usr/local/bin/systemjs-bundle')` is
specified, this is preferred over locally installed `systemjs-bundle`.

However, the best way to build this (using run-within-docker) requires
`systemjs-bundle` to be in your project's
'node_modules/.bin/systemjs-bundle`.

This code makes looking for the different `systemjs-bundle` binaries
very explicit.  The code is written for readability, not performance.